all kind of bugfixing

+++ plugins/org.bonitasoft.studio.connectors.model.edit/src/org/bonitasoft/studio/connector/model/definition/wizard/AbstractDefinitionWizard.java
@@ -138,7 +138,7 @@ public abstract class AbstractDefinitionWizard extends ExtensibleWizard {
             IFolder targetFoler = defStore.getResource() ;
             IFile iconFile = targetFoler.getFile(definitionWorkingCopy.getIcon()) ;
             try{
-                if(iconFile.exists()){
+                if(iconFile.exists() && !iconFile.getLocation().toFile().getAbsolutePath().equals(imageFile.getAbsolutePath())){
                     iconFile.delete(true, Repository.NULL_PROGRESS_MONITOR) ;
                 }
                 BufferedImage image = ImageIO.read(imageFile) ;
